# Conflicts:
#	backend/src/main/java/com/ssafy/cheertogether/game/domain/Game.java

+++ backend/src/main/java/com/ssafy/cheertogether/league/dto/LeagueResponseExceptTeamList.java
@@ -1,9 +1,6 @@
 package com.ssafy.cheertogether.league.dto;
 
-import java.util.List;
-
 import com.ssafy.cheertogether.league.domain.League;
-import com.ssafy.cheertogether.team.domain.Team;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,10 +11,9 @@ public class LeagueResponseExceptTeamList {
 	private String name;
 	private String hanName;
 	private String logo;
-	private int apiId;
+	private Long apiId;
 
-
-	public LeagueResponseExceptTeamList (League league) {
+	public LeagueResponseExceptTeamList(League league) {
 		this.name = league.getName();
 		this.hanName = league.getHanName();
 		this.logo = league.getLogo();
